Clean-up and split patch logic
