+++ src/patch.rs
@@ -1,187 +1,137 @@
-use json_patch::PatchOperation;
-use k8s_openapi::api::core::v1::{Container, PodSpec, Volume, VolumeMount};
+use anyhow::{anyhow, Result};
+use json_patch::{AddOperation, PatchOperation, ReplaceOperation};
+use k8s_openapi::api::core::v1::{Container, EmptyDirVolumeSource, PodSpec, VolumeMount};
 use serde::Serialize;
+use tracing::debug;
 
-#[derive(Clone, Debug, Default)]
-pub struct MakePatch {
-    container_name: String,
-    template_spec: PodSpec,
-    volume: Volume,
-    init_container: Container,
-    mount: VolumeMount,
-    patches: Vec<PatchOperation>,
+pub trait JsonPatch {
+    fn generate_patch(self) -> Result<json_patch::Patch>;
 }
 
-#[async_trait::async_trait]
-trait MakePatcher<T>
-where
-    T: Serialize,
-{
-    type Patcher: Patcher;
-    async fn make_patcher(self, spec: T, comm: String) -> Self::Patcher;
-}
+impl JsonPatch for PodSpec {
+    fn generate_patch(self) -> Result<json_patch::Patch> {
+        let mut patches: Vec<PatchOperation> = vec![];
 
-#[async_trait::async_trait]
-trait Patcher {
-    async fn patch(self) -> anyhow::Result<json_patch::Patch>;
-}
-
-pub struct MakeAwaitPodPatcher;
-#[async_trait::async_trait]
-impl MakePatcher<PodSpec> for MakeAwaitPodPatcher {
-    type Patcher = PodPatcher;
-    async fn make_patcher(self, spec: PodSpec, comm: String) -> PodPatcher {
-        PodPatcher {
-            spec,
-            comm,
-            ..Default::default()
+        if self.init_containers.is_none() {
+            patches.push(mk_add_patch("/spec/initContainers", {}));
         }
-    }
-}
+        patches.push(mk_add_patch(
+            "/spec/initContainers/-",
+            create_curl_container(),
+        ));
 
-#[derive(Debug, Default)]
-struct PodPatcher {
-    spec: PodSpec,
-    comm: String,
-    ops: Vec<PatchOperation>,
-}
+        if self.volumes.is_none() {
+            patches.push(mk_add_patch("/spec/volumes", {}));
+        }
 
-#[async_trait::async_trait]
-impl Patcher for PodPatcher {
-    async fn patch(self) -> anyhow::Result<json_patch::Patch> {
-        todo!();
-    }
-}
-
-impl MakePatch {
-    pub fn new(container_name: String, template_spec: PodSpec) -> Self {
-        Self {
-            container_name,
-            template_spec,
-            patches: vec![],
-            volume: k8s_openapi::api::core::v1::Volume {
-                name: String::from("linkerd-await"),
-                empty_dir: Some(Default::default()),
+        patches.push(mk_add_patch(
+            "/spec/volumes/-",
+            k8s_openapi::api::core::v1::Volume {
+                name: "linkerd-await".into(),
+                empty_dir: Some(EmptyDirVolumeSource::default()),
                 ..Default::default()
             },
-            mount: k8s_openapi::api::core::v1::VolumeMount {
-                mount_path: String::from("/linkerd"),
-                name: String::from("linkerd-await"),
-                ..Default::default()
-            },
-            ..Default::default()
-        }
-    }
+        ));
 
-    pub fn add_await_volume(self) -> Self {
-        let mut patches = self.patches;
-        let vol_len = self
-            .template_spec
-            .volumes
-            .as_ref()
-            .unwrap_or(&Vec::new())
-            .len();
-        if vol_len == 0 {
-            patches.push(PatchOperation::Add(create_patch_add(
-                "/spec/template/spec/volumes",
-                serde_json::json!(Vec::<Volume>::new()),
-            )));
-        }
-        patches.push(json_patch::PatchOperation::Add(create_patch_add(
-            "/spec/template/spec/volumes/-",
-            serde_json::json!(self.volume.clone()),
-        )));
-        Self { patches, ..self }
-    }
-
-    pub fn add_init_container(self) -> Self {
-        let mut patches = self.patches;
-        let init_len = self
-            .template_spec
-            .init_containers
-            .as_ref()
-            .unwrap_or(&Vec::new())
-            .len();
-        // Create patch with path if it doesn't exist
-        if init_len == 0 {
-            patches.push(PatchOperation::Add(create_patch_add(
-                "/spec/template/spec/initContainers",
-                serde_json::json!(Vec::<Container>::new()),
-            )));
-        }
-
-        let init_container = {
-            let mount = self.mount.clone();
-            Container {
-                name: String::from("linkerd-await"),
-                image: Some(String::from("docker.io/matei207/init-await:v0.0.1")),
-                command: Some(vec![String::from("cp")]),
-                args: Some(vec![
-                    String::from("/tmp/linkerd-await"),
-                    String::from("/linkerd-await"),
-                ]),
-                volume_mounts: Some(vec![mount]),
-                ..Default::default()
+        for (i, container) in self.containers.into_iter().enumerate() {
+            let name = container.name.clone();
+            // skip if proxy
+            if name == "linkerd-proxy" {
+                continue;
             }
-        };
-        patches.push(json_patch::PatchOperation::Add(create_patch_add(
-            "/spec/template/spec/initContainers/-",
-            serde_json::json!(init_container),
-        )));
 
-        Self {
-            patches,
-            init_container,
-            ..self
-        }
-    }
-
-    pub fn add_volume_to_container(self) -> anyhow::Result<Self> {
-        let mut patches = self.patches;
-        let container = {
-            let c = if let Some(v) =
-                find_container(&self.container_name, &self.template_spec.containers)
-            {
-                v
-            } else {
-                tracing::error!(container_name=%self.container_name, "container does not exist");
-                anyhow::bail!(
-                    "container {} does not exist in pod template spec",
-                    self.container_name
-                );
-            };
-            let mut mounts = c.volume_mounts.unwrap_or(vec![]);
-            mounts.push(self.mount.clone());
-            Container {
-                volume_mounts: Some(mounts),
-                ..c
+            let current_path = format!("{}/{}", "/spec/containers", i);
+            match create_container_patches(&current_path, container) {
+                Ok(mut container_patches) => {
+                    patches.append(&mut container_patches);
+                    debug!(container_name=%name, "Patched container");
+                }
+                Err(err) => {
+                    debug!(container_name=%name, %err, "Skipped patch");
+                }
             }
-        };
+        }
 
-        patches.push(json_patch::PatchOperation::Add(create_patch_add(
-            "/spec/template/spec/containers/-",
-            serde_json::json!(container),
-        )));
-
-        Ok(Self { patches, ..self })
-    }
-
-    pub fn build_patch(self) -> json_patch::Patch {
-        tracing::debug!(?self.patches, "patches");
-        json_patch::Patch(self.patches)
+        Ok(json_patch::Patch(patches))
     }
 }
 
-fn find_container(name: &str, containers: &Vec<Container>) -> Option<Container> {
-    containers
-        .iter()
-        .find(|container| &container.name == name)
-        .map(|c| c.clone())
+fn create_container_patches(root_path: &str, c: Container) -> Result<Vec<PatchOperation>> {
+    let comm = c
+        .command
+        .as_ref()
+        .ok_or_else(|| anyhow!("container {} is missing 'command' field", c.name))?;
+
+    let mut new_args = vec!["--shutdown".into(), "--".into()];
+    for command in comm.clone().into_iter() {
+        new_args.push(command);
+    }
+
+    if let Some(args) = c.args {
+        for arg in args.into_iter() {
+            new_args.push(arg);
+        }
+    }
+
+    let mut patches = Vec::new();
+    let comm_path = format!("{}/command", root_path);
+    patches.push(mk_replace_patch(comm_path, vec!["/linkerd/linkerd-await"]));
+
+    let arg_path = format!("{}/args", root_path);
+    patches.push(mk_replace_patch(arg_path, new_args));
+
+    if c.volume_mounts.is_none() {
+        let volume_mount_path = format!("{}/volumeMounts", root_path);
+        patches.push(mk_root_patch(volume_mount_path));
+    }
+
+    let volume_path = format!("{}/volumeMounts/-", root_path);
+    patches.push(mk_add_patch(volume_path, create_volume_mount(true)));
+
+    Ok(patches)
 }
 
-fn create_patch_add(path: &str, value: serde_json::Value) -> json_patch::AddOperation {
-    json_patch::AddOperation {
+fn mk_add_patch<T: Serialize, S: Into<String>>(path: S, value: T) -> PatchOperation {
+    PatchOperation::Add(AddOperation {
         path: path.into(),
-        value,
+        value: serde_json::json!(value),
+    })
+}
+
+fn mk_replace_patch<T: Serialize, S: Into<String>>(path: S, value: T) -> PatchOperation {
+    PatchOperation::Replace(ReplaceOperation {
+        path: path.into(),
+        value: serde_json::json!(value),
+    })
+}
+
+fn mk_root_patch<S: Into<String>>(path: S) -> Result<PatchOperation, _> {
+    Ok(PatchOperation::Add(AddOperation {
+        path: path.into(),
+        value: serde_json::json!({}),
+    }))
+}
+
+fn create_volume_mount(read_only: bool) -> VolumeMount {
+    VolumeMount {
+        mount_path: "/linkerd".to_owned(),
+        name: "linkerd-await".to_owned(),
+        read_only: Some(read_only),
+        ..Default::default()
+    }
+}
+
+fn create_curl_container() -> Container {
+    let mut args = vec!["-c".into()];
+    let comm = format!("cp {} {}", "/tmp/linkerd-await", "/linkerd/linkerd-await");
+    args.push(comm);
+    Container {
+        name: "await-init".to_owned(),
+        image: Some("ghcr.io/mateiidavid/await-util:test".to_owned()),
+        image_pull_policy: Some("IfNotPresent".to_owned()),
+        volume_mounts: Some(vec![create_volume_mount(false)]),
+        command: Some(vec!["/bin/sh".into()]),
+        args: Some(args),
+        ..Default::default()
     }
 }
